WD 25 PUBLISH m1

+++ mqtt-sn-codec/src/main/java/org/slj/mqtt/sn/wire/AbstractMqttsnMessage.java
@@ -87,7 +87,7 @@ public abstract class AbstractMqttsnMessage implements IMqttsnMessage {
                 readHeaderByteWithOffset(data, startIdx)) == 1;
     }
 
-    protected int readUInt8Adjusted(byte[] data, int startIdx) {
+    protected short readUInt8Adjusted(byte[] data, int startIdx) {
         return MqttsnWireUtils.read8bit(
                 readHeaderByteWithOffset(data, startIdx));
     }
